main

+++ ExcelParser/src/com/kharevich/util/HttpUtil.java
@@ -10,20 +10,24 @@ import java.net.URLConnection;
 
 public class HttpUtil {
 	public static File download(String uri, String fileName) throws IOException {
-		URL url = null;
-		URLConnection con = null;
-		int i;
-		url = new URL(uri);
-		con = url.openConnection();
 		File file = new File(fileName);
-		BufferedInputStream bis = new BufferedInputStream(con.getInputStream());
-		BufferedOutputStream bos = new BufferedOutputStream(
-				new FileOutputStream(file.getName()));
-		while ((i = bis.read()) != -1) {
-			bos.write(i);
+		if (!file.exists()) {
+			URL url = null;
+			URLConnection con = null;
+			int i;
+			url = new URL(uri);
+			con = url.openConnection();
+
+			BufferedInputStream bis = new BufferedInputStream(
+					con.getInputStream());
+			BufferedOutputStream bos = new BufferedOutputStream(
+					new FileOutputStream(file.getName()));
+			while ((i = bis.read()) != -1) {
+				bos.write(i);
+			}
+			bos.flush();
+			bis.close();
 		}
-		bos.flush();
-		bis.close();
 		return file;
 	}
 }
